Linked To Update Menu Item Page
Had to change the way that I switch to the update menu page due to trying to use a singleton class to store the reference to the menu item that is being updated.

If I kept it the old way then the singleton class will not work as intended due to the browser refreshing.