Place Order Pt1
Added the code for the "order" button to make the first API call to add the order to the order status database table.

I also added logic to check if the API failed and if so then to retry a max of 5 times.

This is done because the error could possibly be due to the order id that is generating matching an already existing order id. So, it will retry 5 times max to try to generate a new order id.